Update Select or cancel a backup folder
