+++ app/src/main/java/com/seafile/seadroid2/ui/activity/BrowserActivity.java
@@ -31,6 +31,7 @@ import android.support.v4.content.LocalBroadcastManager;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.Toolbar;
+import android.text.TextUtils;
 import android.util.Log;
 import android.view.KeyEvent;
 import android.view.Menu;
@@ -48,7 +49,6 @@ import com.seafile.seadroid2.account.Account;
 import com.seafile.seadroid2.account.AccountManager;
 import com.seafile.seadroid2.backupdirectory.FileDirService;
 import com.seafile.seadroid2.backupdirectory.FileDirService.FileDirBinder;
-import com.seafile.seadroid2.backupdirectory.FolderBean;
 import com.seafile.seadroid2.cameraupload.CameraUploadManager;
 import com.seafile.seadroid2.cameraupload.MediaObserverService;
 import com.seafile.seadroid2.data.CheckUploadServiceEvent;
@@ -105,7 +105,6 @@ import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
 import org.greenrobot.eventbus.ThreadMode;
 import org.json.JSONException;
-import org.litepal.LitePal;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -832,20 +831,10 @@ public class BrowserActivity extends BaseActivity
             dirService = binder.getService();
             Log.d(DEBUG_TAG, "-----bind FileDirService");
             boolean dirAutomaticUpload = SettingsManager.instance().isDirAutomaticUpload();
-            if (dirAutomaticUpload) {
-                List<FolderBean> paths = LitePal.findAll(FolderBean.class);
-                if (dirService != null && paths != null) {
-                    Utils.utilsLogInfo(false, "--dirService-------Data-");
-                    for (FolderBean fp : paths) {
-                        if (fp.getEmail().equals(account.email)) {
-                            dirService.uploadFile(account, fp);
-                        }
-                    }
-
-                }
+            String backupEmail = SettingsManager.instance().getBackupEmail();
+            if (dirAutomaticUpload && dirService != null && !TextUtils.isEmpty(backupEmail)) {
+                dirService.uploadFile(backupEmail);
             }
-
-
         }
 
         @Override
