--- app/src/main/res/layout/dir_welcome_fragment.xml
@@ -1,40 +1,0 @@
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent">
-
-    <TextView
-        android:id="@+id/cuc_wel_title_tv"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentLeft="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginLeft="@dimen/tv_title_margin_left"
-        android:layout_marginTop="@dimen/tv_title_margin_top"
-        android:textColor="@color/fancy_orange"
-        android:text="@string/settings_cuc_title"
-        android:textSize="@dimen/tv_title_txt_size" />
-
-    <TextView
-        android:id="@+id/cuc_wel_txt_tv"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignLeft="@+id/cuc_wel_title_tv"
-        android:layout_below="@+id/cuc_wel_title_tv"
-        android:layout_marginTop="@dimen/section_margin_top"
-        android:text="@string/settings_dir_welcome_text"
-        android:textColor="@color/fancy_gray"
-        android:layout_marginRight="30dp"
-        android:textSize="@dimen/tv_subtitle_txt_size" />
-
-    <TextView
-            android:id="@+id/cuc_swipe_left_txt_tv"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignLeft="@+id/cuc_wel_txt_tv"
-            android:layout_below="@+id/cuc_wel_txt_tv"
-            android:layout_marginTop="@dimen/tv_subtitle_margin_top"
-            android:text="@string/settings_cuc_swipe_left"
-            android:textColor="@color/fancy_gray"
-            android:textSize="@dimen/tv_subtitle_txt_size" />
-
-</RelativeLayout>
