+++ app/src/main/res/values/strings.xml
@@ -588,10 +588,16 @@
   <string name="folder_backup_completed">The folder is backed up</string>
   <string name="get_storage_permission_failed">Failed to get storage permission</string>
   <string name="authorization_storage_permission">Authorization denied permanently, please grant storage permission manually.</string>
-  <string name="settings_dir_welcome_text">Thanks for using the directory upload configuration helper. \nTo help you get started with the folder backup service, you\'ll be guided through a few initial steps.</string>
   <string name="settings_folder_upload_repo_hint">0</string>
   <string name="settings_folder_upload_change_dir_title">The selected backup directory</string>
-  <string name="folder_waiting_state">Folder upload waiting</string>
-  <string name="folder_upload_state">Folder upload Info</string>
+  <string name="settings_folder_upload_traffic">Backup mode</string>
+  <string name="settings_folder_upload_traffic_hint">wifi</string>
+  <string name="settings_folder_upload_library">Library backup to</string>
+  <string name="folder_waiting_state">Folder backup waiting</string>
+  <string name="folder_upload_state">Folder backup info</string>
+  <string name="folder_backup_mode">WIFI or mobile traffic</string>
+  <string name="folder_backup_mode_title">Choose a mode</string>
+  <string name="folder_backup_select_repo_hint">Please select</string>
+  <string name="folder_backup_select_repo_update">Update backup library</string>
 
 </resources>
